feat: refactor of CheckConsistency function to support schematics
