+++ testhelper/terraform.go
@@ -1,10 +1,9 @@
 package testhelper
 
 import (
+	"encoding/json"
 	"errors"
 	"fmt"
-	"github.com/gruntwork-io/terratest/modules/files"
-	"github.com/terraform-ibm-modules/ibmcloud-terratest-wrapper/common"
 	"log"
 	"os"
 	"os/exec"
@@ -12,6 +11,13 @@ import (
 	"reflect"
 	"regexp"
 	"strings"
+
+	"github.com/gruntwork-io/terratest/modules/files"
+	"github.com/gruntwork-io/terratest/modules/logger"
+	"github.com/gruntwork-io/terratest/modules/terraform"
+	tfjson "github.com/hashicorp/terraform-json"
+	"github.com/stretchr/testify/assert"
+	"github.com/terraform-ibm-modules/ibmcloud-terratest-wrapper/common"
 )
 
 // RemoveFromStateFile Attempts to remove resource from state file
@@ -63,9 +69,9 @@ func ValidateTerraformOutputs(outputs map[string]interface{}, expectedKeys ...st
 		if !ok {
 			missingKeys = append(missingKeys, key)
 			if err != nil {
-				err = fmt.Errorf("%wOutput: %s'%s'%s was not found\n", err, blueBold, key, reset)
+				err = fmt.Errorf("%wOutput: %s'%s'%s was not found", err, blueBold, key, reset)
 			} else {
-				err = fmt.Errorf("Output: %s'%s'%s was not found\n", blueBold, key, reset)
+				err = fmt.Errorf("output: %s'%s'%s was not found", blueBold, key, reset)
 			}
 		} else {
 			if value == nil || (reflect.TypeOf(value).String() == "string" && len(strings.Trim(value.(string), " ")) == 0) {
@@ -77,9 +83,9 @@ func ValidateTerraformOutputs(outputs map[string]interface{}, expectedKeys ...st
 					expected = "blank string"
 				}
 				if err != nil {
-					err = fmt.Errorf("%wOutput: %s'%s'%s was not expected to be %s\n", err, blueBold, key, reset, expected)
+					err = fmt.Errorf("%wOutput: %s'%s'%s was not expected to be %s", err, blueBold, key, reset, expected)
 				} else {
-					err = fmt.Errorf("Output: %s'%s'%s was not expected to be %s\n", blueBold, key, reset, expected)
+					err = fmt.Errorf("output: %s'%s'%s was not expected to be %s", blueBold, key, reset, expected)
 				}
 			}
 		}
@@ -119,5 +125,146 @@ func CleanTerraformDir(directory string) {
 				log.Printf("Error removing file %s: %s", fileName, err)
 			}
 		}
+	}
+}
+
+// checkConsistency Fails the test if any destroys are detected and the resource is not exempt.
+// If any addresses are provided in IgnoreUpdates.List then fail on updates too unless the resource is exempt
+// Returns TRUE if there were consistency changes that were identified
+func CheckConsistency(plan *terraform.PlanStruct, testOptions CheckConsistencyOptionsI) bool {
+	validChange := false
+
+	// extract consistency options from base set of options (schematic or terratest)
+	options := testOptions.GetCheckConsistencyOptions()
+
+	for _, resource := range plan.ResourceChangesMap {
+		// get JSON string of full changes for the logs
+		changesBytes, changesErr := json.MarshalIndent(resource.Change, "", "  ")
+		// if it errors in the marshall step, just put a placeholder and move on, not important
+		changesJson := "--UNAVAILABLE--"
+		if changesErr == nil {
+			changesJson = string(changesBytes)
+		}
+
+		var resourceDetails string
+
+		// Treat all keys in the BeforeSensitive and AfterSensitive maps as sensitive
+		// Assuming BeforeSensitive and AfterSensitive are of type interface{}
+		beforeSensitive, beforeSensitiveOK := resource.Change.BeforeSensitive.(map[string]interface{})
+		afterSensitive, afterSensitiveOK := resource.Change.AfterSensitive.(map[string]interface{})
+
+		// Create the mergedSensitive map
+		mergedSensitive := make(map[string]interface{})
+
+		// Check if BeforeSensitive is of the expected type
+		if beforeSensitiveOK {
+			// Copy the keys and values from BeforeSensitive to the mergedSensitive map.
+			for key, value := range beforeSensitive {
+				mergedSensitive[key] = value
+			}
+		}
+
+		// Check if AfterSensitive is of the expected type
+		if afterSensitiveOK {
+			// Copy the keys and values from AfterSensitive to the mergedSensitive map.
+			for key, value := range afterSensitive {
+				mergedSensitive[key] = value
+			}
+		}
+
+		// Perform sanitization
+		sanitizedChangesJson, err := sanitizeResourceChanges(resource.Change, mergedSensitive)
+		if err != nil {
+			sanitizedChangesJson = "Error sanitizing sensitive data"
+			logger.Log(options.Testing, sanitizedChangesJson)
+		}
+		formatChangesJson, err := common.FormatJsonStringPretty(sanitizedChangesJson)
+
+		var formatChangesJsonString string
+		if err != nil {
+			logger.Log(options.Testing, "Error formatting JSON, use unformatted")
+			formatChangesJsonString = sanitizedChangesJson
+		} else {
+			formatChangesJsonString = string(formatChangesJson)
+		}
+
+		diff, diffErr := common.GetBeforeAfterDiff(changesJson)
+
+		if diffErr != nil {
+			diff = fmt.Sprintf("Error getting diff: %s", diffErr)
+		} else {
+			// Split the changesJson into "Before" and "After" parts
+			beforeAfter := strings.Split(diff, "After: ")
+
+			// Perform sanitization on "After" part
+			var after string
+			if len(beforeAfter) > 1 {
+				after, err = common.SanitizeSensitiveData(beforeAfter[1], mergedSensitive)
+				handleSanitizationError(err, "after diff", options)
+			} else {
+				after = "Could not parse after from diff" // dont print incase diff contains sensitive values
+			}
+
+			// Perform sanitization on "Before" part
+			var before string
+			if len(beforeAfter) > 0 {
+				before, err = common.SanitizeSensitiveData(strings.TrimPrefix(beforeAfter[0], "Before: "), mergedSensitive)
+				handleSanitizationError(err, "before diff", options)
+			} else {
+				before = "Could not parse before from diff" // dont print incase diff contains sensitive values
+			}
+
+			// Reassemble the sanitized diff string
+			diff = "  Before: \n\t" + before + "\n  After: \n\t" + after
+		}
+		resourceDetails = fmt.Sprintf("\nName: %s\nAddress: %s\nActions: %s\nDIFF:\n%s\n\nChange Detail:\n%s", resource.Name, resource.Address, resource.Change.Actions, diff, formatChangesJsonString)
+
+		var errorMessage string
+		if !options.IgnoreDestroys.IsExemptedResource(resource.Address) {
+			errorMessage = fmt.Sprintf("Resource(s) identified to be destroyed %s", resourceDetails)
+			assert.False(options.Testing, resource.Change.Actions.Delete(), errorMessage)
+			assert.False(options.Testing, resource.Change.Actions.DestroyBeforeCreate(), errorMessage)
+			assert.False(options.Testing, resource.Change.Actions.CreateBeforeDestroy(), errorMessage)
+			validChange = true
+		}
+		if !options.IgnoreUpdates.IsExemptedResource(resource.Address) {
+			errorMessage = fmt.Sprintf("Resource(s) identified to be updated %s", resourceDetails)
+			assert.False(options.Testing, resource.Change.Actions.Update(), errorMessage)
+			validChange = true
+		}
+		// We only want to check pure Adds (creates without destroy) if the consistency test is
+		// NOT the result of an Upgrade, as some adds are expected when doing the Upgrade test
+		// (such as new resources were added as part of the pull request)
+		if !options.IsUpgradeTest {
+			if !options.IgnoreAdds.IsExemptedResource(resource.Address) {
+				errorMessage = fmt.Sprintf("Resource(s) identified to be created %s", resourceDetails)
+				assert.False(options.Testing, resource.Change.Actions.Create(), errorMessage)
+				validChange = true
+			}
+		}
+	}
+
+	return validChange
+}
+
+// sanitizeResourceChanges sanitizes the sensitive data in a Terraform JSON Change and returns the sanitized JSON.
+func sanitizeResourceChanges(change *tfjson.Change, mergedSensitive map[string]interface{}) (string, error) {
+	// Marshal the Change to JSON bytes
+	changesBytes, err := json.MarshalIndent(change, "", "  ")
+	if err != nil {
+		return "", err
+	}
+	changesJson := string(changesBytes)
+
+	// Perform sanitization of sensitive data
+	changesJson, err = common.SanitizeSensitiveData(changesJson, mergedSensitive)
+	return changesJson, err
+}
+
+// handleSanitizationError logs an error message if a sanitization error occurs.
+func handleSanitizationError(err error, location string, options *CheckConsistencyOptions) {
+	if err != nil {
+		errorMessage := fmt.Sprintf("Error sanitizing sensitive data in %s", location)
+		logger.Log(options.Testing, errorMessage)
 	}
 }
